refactor(core): extract RowState into signal-list module
