+++ src/frontend/packages/core/src/shared/components/signal-list/row-state.types.ts
@@ -1,0 +1,23 @@
+export interface RowsState {
+  [rowUID: string]: RowState;
+}
+
+export interface RowState {
+  busy?: boolean;
+  error?: boolean;
+  message?: string;
+  blocked?: boolean;
+  highlighted?: boolean;
+  deleting?: boolean;
+  warning?: boolean;
+  disabled?: boolean;
+  [customState: string]: any;
+}
+
+export const getDefaultRowState = (): RowState => ({
+  busy: false,
+  error: false,
+  blocked: false,
+  deleting: false,
+  message: null
+});
